Rudimentary reputation Very simple for now: Assume a False POsitive Rate (hardcoded for now), measure the z-score (distance from expected in Standard Deviation), assume greater than 1 std-dev should be validated for every inference, lower only 1 in 10.
Also add a Ramping for new Participants, since we should be checking every inference until they reach some minimum.

I derive the minimum needed from the FPR, using the minimum value that a single failure would be considered OK as a guess. So, the lower the FPR, the more inferences needed to say they're good to go. Not perfect, but better than an arbitrary number