Cosmos 0.53.3 upgrade (plus Replay attacks) (#205)
* Finishing the implementation, still tests to fix

* Minor review fixes

* For now, ALL signatures based on the original, unmodified prompt.
That means we have to store BOTH prompts (one to repro correctly, one to verify signature). It's not ideal, especially if the payloads get really large, as they are wont to do in modern AI. We should look into optimizations in the future.

* Fix consumer tests

* Fix streaming tests

* A little refactoring

* WIP commit

* Move to released version

* Missing file

* Add needed upgrade code.
The big thing is to call RunMigrations (which we should have been). But that caused issues with the `capability` module, as it seems to have data on the chain despite having no module version logged.

So, we fake that it already exists. It works, but I'm still not super happy about how.

* Fix tests and unit tests

* Added config parameters for the timeouts for our Inference requests/nonces

* Minor fixes

---------

Co-authored-by: Gleb Morgachev <morgachev.g@gmail.com>